requirements.txt
ircchallenge ~=0.1.2
ircrobots    ~=0.6.6
pyyaml       ~=6.0.1
